requirements.txt
groq
fastapi
python-dotenv
